cbc_mac_breaker/message_forgery1/backend/requirements.txt
Werkzeug==2.0.1
Flask==2.0.1
PyCryptodome==3.11.0